requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
pydantic>=2.11,<3.0
python-dotenv>=1.0,<2.0
earthengine-api>=1.5,<2.0
cdsapi>=0.7.7,<1.0
netCDF4>=1.7,<2.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
